requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-environ==0.4.5
django-extensions==3.1.0
django-model-utils==4.1.0
django-rest-registration==0.5.6
django-timezone-field==4.1.1
djangorestframework==3.12.2
drf-yasg==1.20.0
graphviz==0.15
idna==2.10
importlib-metadata==3.1.0
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.7
prompt-toolkit==3.0.8
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
stripe==2.55.0
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0
